Moved all of the getters [demos][skip hazardlib]
