+++ openquake/commands/plot_hmaps.py
@@ -18,6 +18,7 @@
 
 from __future__ import print_function
 from openquake.baselib import sap, datastore
+from openquake.calculators import getters
 from openquake.commonlib import calc
 
 
@@ -50,7 +51,7 @@ def plot_hmaps(calc_id):
     """
     dstore = datastore.read(calc_id)
     oq = dstore['oqparam']
-    mean = calc.PmapGetter(dstore).get_mean()
+    mean = getters.PmapGetter(dstore).get_mean()
     hmaps = calc.make_hmap(mean, oq.imtls, oq.poes)
     M, P = len(oq.imtls), len(oq.poes)
     array = hmaps.array.reshape(len(hmaps.array), M, P)
